elia_noises_offset_finder/requirements.txt
audioread==3.0.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
decorator==5.2.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
joblib==1.5.1
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.44.0
matplotlib==3.10.0
moviepy==2.2.1
msgpack==1.1.1
numba==0.61.2
numpy==2.2.6
ocrmac==1.0.0
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
proglog==0.1.12
pycparser==2.22
pyobjc-core==11.1
pyobjc-framework-cocoa==11.1
pyobjc-framework-coreml==11.1
pyobjc-framework-quartz==11.1
pyobjc-framework-vision==11.1
python-dotenv==1.1.0
requests==2.32.4
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-extensions==4.14.0
urllib3==2.4.0
